feat: Cron: routine
While the API is on, daily at 3 am an asynchronous function will run that updates the Launchers API data.
The early morning time was determined to be a time with a possibility of less traffic.
Now the FillContext function is being called properly when starting the API through the Scheduler.
Scheduler class created to centralize functions related to the routines' behavior.